change EXPECT to be 'lower/greater than' rather than 'equal'
This is because the minimum gain in the stop band of a filter,
is not know, just has to be 'low enough'. Similarly,
gain in the passing-band has to be 'high enough'.
At least for being a band-pass filter, when we offer a way to control
the gain of the filter, we will add a specific test for it.