added interface for signal analysis
This is an interface fde define modules that only analysis signals
but do not perform synthesis. Same as shorttimeprocess but
the sythesis frame function is empty.

It is good for algorithms that just extract information from the signal
and take decision based on the parameterisations. Such as sound localisation,
acoustic event classifications, etc.