now shorttime process inherits from shorttimeprocessor
The processor interface is a simplification of the shorttimeprocess
interface. This higer level interface is intendet for its use
in external libraries that do not need to know much details about
the processor.

They just need to know that the module processes buffer signal. This is
way the interface almost only define the process fuctions plus a few helpers