added MedianFilter
Median filter is a filter that replaces each input value
by the median value obtain at a maks_size window around it.
It produces a delay that is equal to half the mask size.
It keeps memory by means of a circular buffer and calculates future values
depending on past ones.