changed BankFilter API, to correct the definition of BankFilter
A bank filter, consists of a set of filters that is applied to a signal, one by one.
It is not that each filter is applied to the output of the previous one.
The input signal is used to feed each and every of the filters and the outcome
of a filter bank is a vector of signals that represent the output of every filter in
the filter bank.
Previous implemenetation was wrong, so I had to refactor the API here, to not deliver
a residual signal, because it has no sense here.